la gestion des indemnites l'ors de la suppression des abonnements

+++ src/Service/AbonnementServices/AbonnementNormalService.php
@@ -239,7 +239,17 @@ class AbonnementNormalService{
             // Si l'abonnement est déjà payé, alors on ne peut pas le supprimer.
             if($abonnementObject->getMontantPayee()) return "L'abonnement est déjà payé, vous ne pouvez donc pas le supprimer." ;
             //////////////////////////////////////////////////////////////////
-            
+ 
+            // Avant de supprimer un abonnement, il faut s'assurer que les indemnités des formateurs associées à cet abonnement ne sont pas encore payées
+            $indemnitesDeFormateures = $this->obtenirIndemnitesDeFormateuresService->obtenirIndemnitesParAbonnement($abonnementObject);
+            if( $indemnitesDeFormateures === false) return 'Une erreur est survenue.' ;
+            foreach ( $indemnitesDeFormateures as $indemnite){
+                if($indemnite->getMontantPayee()) return "Vous ne pouvez pas supprimer cet abonnement car vous avez déjà payé un formateur en se basant sur cet abonnement" ;
+            }
+            // Si tout est OK, l'abonnement sera supprimé et les indemnités des formateurs seront supprimées automatiquement, 
+            // car dans l'entité "IndemnitesDeFormateurs" nous avons utilisé <<onDelete: 'CASCADE'>>
+            ////////////////////////
+
             $this->entityManager->remove($abonnementObject);
             $this->entityManager->flush();
